pyproject.toml
[tool.poetry]
name = "app1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.39.0"
langchain = "^0.3.4"
openai = "^1.52.2"
python-docx = "^1.1.2"
pymupdf = "^1.24.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
